+++ packages/inqvine_core_firebase/lib/inqvine_core_firebase.dart
@@ -1,6 +1,7 @@
 library inqvine_core_firebase;
 
 export 'services/inqvine_auth_service.dart';
+export 'services/inqvine_admin_service.dart';
 
 export 'package:firebase_auth/firebase_auth.dart';
 export 'package:cloud_firestore/cloud_firestore.dart';
@@ -10,6 +11,7 @@ import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:firebase_crashlytics/firebase_crashlytics.dart';
 
+import 'package:inqvine_core_firebase/services/inqvine_admin_service.dart';
 import 'package:inqvine_core_firebase/services/inqvine_auth_service.dart';
 import 'package:inqvine_core_main/inqvine_core_main.dart';
 
@@ -22,6 +24,7 @@ class InqvineCoreFirebase {
 
   Future<void> registerInqvineFirebaseServices() async {
     await inqvine.registerService<InqvineAuthService>(authService);
+    await inqvine.registerService<InqvineAdminService>(adminService);
 
     inqvine.registerInLocator<FirebaseAuth>(FirebaseAuth.instance);
     inqvine.registerInLocator<FirebaseFirestore>(FirebaseFirestore.instance);
@@ -33,4 +36,5 @@ class InqvineCoreFirebase {
   FirebaseCrashlytics get firebaseCrashlytics => inqvine.getFromLocator<FirebaseCrashlytics>();
 
   InqvineAuthService get authService => inqvine.getFromLocator<InqvineAuthService>();
+  InqvineAdminService get adminService => inqvine.getFromLocator<InqvineAdminService>();
 }