+++ packages/inqvine_core_main/CHANGELOG.md
@@ -1,3 +1,7 @@
+## [0.0.5-nullsafety.0] - (InqvineCoreFirebase) InqvineAdminService
+
+* Allow configuration of `admin_users` via a collection, and some firestore rules.
+
 ## [0.0.4-nullsafety.0] - Initial implementation of Inqvine Firebase Core
 
 * Adds auth service. (Note: Only supports email and password login at the moment)